Allow use of zope.viewlet and z3c.template
The zope.viewlet is a big win for factoring templates and reusing
pieces; all of the work to support archiveindex.tmpl gets *much*
cleaner with this and I imagine other things will too. z3c.template
just provides us with a standard interface and ZCML for registering
templates when we don't need an actual file or autodiscovery.

After doing a bunch of work this morning on Chameleon itself, I set
out this morning to document the way the current template system was
working and realized it was sub-par, so this is the result.